#1071 - Move tx tests to .test folder and adjust modules
