+++ transaction/api/src/main/java/module-info.java
@@ -33,5 +33,4 @@ module cloud.piranha.transaction.api {
     requires java.transaction.xa;
 
     exports javax.transaction;
-    opens javax.transaction;
 }
