+++ transaction/api/src/test/java/javax/transaction/test/SystemExceptionTest.java
@@ -24,33 +24,46 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package javax.transaction;
+package javax.transaction.test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
+
+import javax.transaction.SystemException;
+
 import org.junit.jupiter.api.Test;
 
 /**
- * The JUnit tests for the TransactionRolledbackException class.
+ * The JUnit tests for the SystemException class.
  * 
  * @author Manfred Riem (mriem@manorrock.com)
  */
-class TransactionRollbackExceptionTest {
+class SystemExceptionTest {
     
     /**
      * Test getMessage method.
      */
     @Test
     void testGetMessage() {
-        TransactionRolledbackException exception = new TransactionRolledbackException("message");
+        SystemException exception = new SystemException("message");
         assertEquals("message", exception.getMessage());
-    }    
+    }  
+    
     /**
      * Test getMessage method.
      */
     @Test
     void testGetMessage2() {
-        TransactionRolledbackException exception = new TransactionRolledbackException();
+        SystemException exception = new SystemException();
         assertNull(exception.getMessage());
+    }  
+    
+    /**
+     * Test getting error code.
+     */
+    @Test
+    void testGettingErrorCode() {
+        SystemException exception = new SystemException(1);
+        assertEquals(1, exception.errorCode);
     }
 }
