+++ transaction/api/src/test/java/javax/transaction/test/TestTransactional.java
@@ -1,20 +1,20 @@
 /*
  * Copyright (c) 2002-2020 Manorrock.com. All Rights Reserved.
  *
- * Redistribution and use in source and binary forms, with or without 
+ * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
- *   1. Redistributions of source code must retain the above copyright notice, 
+ *   1. Redistributions of source code must retain the above copyright notice,
  *      this list of conditions and the following disclaimer.
  *   2. Redistributions in binary form must reproduce the above copyright notice,
  *      this list of conditions and the following disclaimer in the documentation
  *      and/or other materials provided with the distribution.
- *   3. Neither the name of the copyright holder nor the names of its 
+ *   3. Neither the name of the copyright holder nor the names of its
  *      contributors may be used to endorse or promote products derived from this
  *      software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
  * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
  * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
@@ -24,33 +24,56 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package javax.transaction;
+package javax.transaction.test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import org.junit.jupiter.api.Test;
+import java.lang.annotation.Annotation;
+
+import javax.transaction.Transactional;
 
 /**
- * The JUnit tests for the HeuristicCommitException class.
- * 
+ * A test Transactional.
+ *
  * @author Manfred Riem (mriem@manorrock.com)
  */
-class HeuristicCommitExceptionTest {
-    
+public class TestTransactional implements Transactional {
+
     /**
-     * Test getMessage method.
+     * Don't rollback on.
+     *
+     * @return null.
      */
-    @Test
-    void testGetMessage() {
-        HeuristicCommitException exception = new HeuristicCommitException("message");
-        assertEquals("message", exception.getMessage());
-    }    
+    @Override
+    public Class[] dontRollbackOn() {
+        return null;
+    }
+
     /**
-     * Test getMessage method.
+     * Roll back on.
+     *
+     * @return null.
      */
-    @Test
-    void testGetMessage2() {
-        HeuristicCommitException exception = new HeuristicCommitException();
-        assertNull(exception.getMessage());
+    @Override
+    public Class[] rollbackOn() {
+        return null;
+    }
+
+    /**
+     * Value.
+     *
+     * @return TxType.MANDATORY.
+     */
+    @Override
+    public TxType value() {
+        return TxType.MANDATORY;
+    }
+
+    /**
+     * Annotation type.
+     *
+     * @return the annotation type.
+     */
+    @Override
+    public Class<? extends Annotation> annotationType() {
+        return null;
     }
 }
