+++ transaction/api/src/test/java/javax/transaction/test/module-info.java
@@ -1,20 +1,20 @@
 /*
  * Copyright (c) 2002-2020 Manorrock.com. All Rights Reserved.
  *
- * Redistribution and use in source and binary forms, with or without 
+ * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
- *   1. Redistributions of source code must retain the above copyright notice, 
+ *   1. Redistributions of source code must retain the above copyright notice,
  *      this list of conditions and the following disclaimer.
  *   2. Redistributions in binary form must reproduce the above copyright notice,
  *      this list of conditions and the following disclaimer in the documentation
  *      and/or other materials provided with the distribution.
- *   3. Neither the name of the copyright holder nor the names of its 
+ *   3. Neither the name of the copyright holder nor the names of its
  *      contributors may be used to endorse or promote products derived from this
  *      software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
  * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
  * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
@@ -24,34 +24,18 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package javax.transaction;
 
-import org.junit.jupiter.api.Test;
+module cloud.piranha.transaction.api.test {
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+    requires cloud.piranha.transaction.api;
 
-/**
- * The JUnit tests for the InvalidTransactionException class.
- * 
- * @author Manfred Riem (mriem@manorrock.com)
- */
-class InvalidTransactionExceptionTest {
-    
-    /**
-     * Test getMessage method.
-     */
-    @Test
-    void testGetMessage() {
-        InvalidTransactionException exception = new InvalidTransactionException("message");
-        assertEquals("message", exception.getMessage());
-    }    
-    /**
-     * Test getMessage method.
-     */
-    @Test
-    void testGetMessage2() {
-        InvalidTransactionException exception = new InvalidTransactionException();
-        assertNull(exception.getMessage());
-    }
+    requires jakarta.enterprise.cdi.api;
+    requires jakarta.interceptor.api;
+
+    requires java.rmi;
+    requires java.transaction.xa;
+    requires org.junit.jupiter.api;
+
+    exports javax.transaction.test;
+    opens javax.transaction.test;
 }
